% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldbod2.R
\name{ldbod2}
\alias{ldbod2}
\title{Local Density-Based Outlier Detection using reference data set with flexible nearest neighbor search}
\usage{
ldbod2(X = data, Y = ref.data, k = c(10, 20), method = c("lof", "ldf",
  "rkof", "lpdf"), ldf.param = c(h = 1, c = 0.1), rkof.param = c(alpha = 1,
  C = 1, sig2 = 1), lpdf.param = c(cov.type = "full", sigma2 = 1e-05, tmax =
  1, v = 1), treetype = "kd", searchtype = "standard", eps = 0,
  scale.data = T)
}
\arguments{
\item{X}{An n x p data matrix to compute outlier scores}

\item{Y}{An m x p reference data matrix.}

\item{k}{Neighborhood size, k must be less than m.}

\item{method}{Character vector specifying the local density-based method(s) to compute. User can specify more than
one method.  By default all methods are computed}

\item{ldf.param}{Vector of parameters for method LDF, default values are h=1 and c=0.1}

\item{rkof.param}{Vector  parameters for method RKOF, default values are alpha=1, C=1, sig2=1}

\item{lpdf.param}{Vector of paramters for method LPDF, default values are tmax=1, sigma2=1e-5, and v=1}

\item{treetype}{Character vector specifiying tree method.  Either 'kd' or 'bd' tree may be specified.  Default is 'kd'. Refer to documentation for RANN package.
treetype Character vector specifiying tree method.  Either "kd" or '"d" tree may be specified.  Default is "kd". Refer to documentation for RANN package.}

\item{searchtype}{Character vector specifiying kNN search type. Default value is "standard". Refer to documentation for RANN package.}

\item{eps}{Error bound.  Default is 0.0 which implies exact nearest neighgour search.  Refer to documentation for RANN package.}

\item{scale.data}{Logical value indicating to scale each feature of X using standard noramlization with mean 0 and standard deviation of 1}
}
\value{
A list of length 9 with the elements:

lrd --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value.  Smaller values indicate a point in more outlying.

lof   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Larger values indicate a point in more outlying.

lde   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Smaller values indicate a point in more outlying.

ldf   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Larger values indicate a point in more outlying.

kde   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Smaller values indicate a point in more outlying.

rkof   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Larger values indicate a point in more outlying.

lpde   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Smaller values indicate a point in more outlying.

lpdf  --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Smaller values indicate a point in more outlying.

lpdr   --An n x length(k) matrix where each column vector represents outlier scores for each specifed k value. Smaller values indicate a point in more outlying.

If a method is not specified then returns NULL
}
\details{
Computes the local density-based outlier scores for input data X referencing Y.  Y could be a set of normal points for semi-supervised
outlier detection, or any other set of reference points.  Allows users the flexibility to reference other data sets besides subsets of X.

Four different methods can be implemented LOF, LDF, RKOF, and LPDF.  Each method specified returns densities and relative densities.
Methods LDF and RKOF utilize guassian kernels.

All kNN distances are carried out using the RANN package. Multivariate t densities for method LPDF are
computed using mnormt package.  Refer to specific packages for more details.
}
\examples{
# 500 x 2 data matrix
X <- matrix(rnorm(1000),500,2)

# five outliers
outliers <- matrix(c(rnorm(2,20),rnorm(2,-12),rnorm(2,-8),rnorm(2,-5),rnorm(2,9)),5,2)
 X <- rbind(X,outliers)

# compute outlier scores without subsampling for all methods
scores <- ldbod(X, k=50)

head(scores$lrd); head(scores$rkof)

# plot data and highlight top 5 outliers retured by lof
plot(X)
points(X[order(scores$lof,decreasing=T)[1:5],],col=2)

# plot data and highlight top 5 outliers retured by outlier score lpde
plot(X)
points(X[order(scores$lpde,decreasing=F)[1:5],],col=2)


 # compute outlier scores for k= 10,20 with 10\% subsampling for methods 'lof' and 'lpdf'
scores <- ldbod(X, k = c(10,20),n_sub = 0.10*nrow(X), method = c('lof','lpdf'))

# plot data and highlight top 5 outliers retuned by lof for k=20
plot(X)
points(X[order(scores$lof[,2],decreasing=T)[1:5],],col=2)


}
\references{
M. M. Breunig, H-P. Kriegel, R.T. Ng, and J. Sander (2000). LOF: Identifying density-based local outliers.  In Proc. of ACM
International Conference on Knowledge Discovery and Data Mining, 93-104.

L. J. Latecki, A. Lazarevic, and D. Pokrajac (2007). Outlier Detection with kernel density funcions.  In Proc. of Machine Learning and Data
Mining in Pattern Recognition, 61-75

J. Gao, W. Hu, Z. Zhang, X. Zhang, and O. Wu (2011). RKOF: Robust kernel-based local outlier detection. In Proc. of Advances in Knowledge Discovery and
Data Mining, 270-283.

K. Williams (2016).  Local parametric density-based outlier deteciton and ensemble learning with application to malware detection. PhD dissertation,
University of Texas at San Antonio.
}

